+++ src/json.cpp
@@ -837,7 +837,7 @@ void JSONWriter::arg(const char* name, const char* value, int quotes)
     }
 }
 
-void JSONWriter::arg(const char* name, handle h, int len)
+void JSONWriter::arg(const char* name, handle h, size_t len)
 {
     char buf[16];
 
@@ -851,10 +851,9 @@ void JSONWriter::arg(const char* name, NodeHandle h)
     arg(name, h.as8byte(), 6);
 }
 
-
-void JSONWriter::arg(const char* name, const byte* value, int len)
+void JSONWriter::arg(const char* name, const byte* value, size_t len)
 {
-    char* buf = new char[static_cast<size_t>(len * 4 / 3 + 4)];
+    char* buf = new char[len * 4 / 3 + 4];
 
     Base64::btoa(value, len, buf);
 
@@ -865,12 +864,12 @@ void JSONWriter::arg(const char* name, const byte* value, int len)
 
 void JSONWriter::arg_B64(const char* n, const string& data)
 {
-    arg(n, (const byte*)data.data(), int(data.size()));
+    arg(n, (const byte*)data.data(), data.size());
 }
 
 void JSONWriter::arg_fsfp(const char* n, std::uint64_t fp)
 {
-    arg(n, (const byte*)&fp, int(sizeof(fp)));
+    arg(n, (const byte*)&fp, sizeof(fp));
 }
 
 void JSONWriter::arg_stringWithEscapes(const char* name, const string& value, int quote)
@@ -960,7 +959,7 @@ void JSONWriter::element(int n)
     mJson.append(std::to_string(n));
 }
 
-void JSONWriter::element(handle h, int len)
+void JSONWriter::element(handle h, size_t len)
 {
     char buf[16];
 
@@ -971,14 +970,14 @@ void JSONWriter::element(handle h, int len)
     mJson.append("\"");
 }
 
-void JSONWriter::element(const byte* data, int len)
+void JSONWriter::element(const byte* data, size_t len)
 {
-    char* buf = new char[static_cast<size_t>(len * 4 / 3 + 4)];
+    char* buf = new char[len * 4 / 3 + 4];
 
     len = Base64::btoa(data, len, buf);
 
     mJson.append(elements() ? ",\"" : "\"");
-    mJson.append(buf, static_cast<size_t>(len));
+    mJson.append(buf, len);
 
     delete[] buf;
 
@@ -999,7 +998,7 @@ void JSONWriter::element(const string& data)
 
 void JSONWriter::element_B64(const string& s)
 {
-    element((const byte*)s.data(), int(s.size()));
+    element((const byte*)s.data(), s.size());
 }
 
 void JSONWriter::openobject()
